Various bug fixes and quality improvements (#167)
List of changes:

1. New config arguments (especially in hash grid nerf) which improve
quality significantly
2. Adaptive ray batching (improves nerf quality a lot)
3. `bg_color` is now a tuple `(0.0, 0.0, 0.0)` not a string
4. Supporting other validation splits `--valid-split`
5. `apex` based fused Adam
6. Bug fix in rays for `dist_min`, inproper normalization in
`channel_fn`
7. 2D support for hash grid
8. `MultiTable` utility class for working with hash grids (and similar)
easier
9. App renderer will now try to make FPS stable for better UX 
10. Fixing iteration count
11. Fix tracker's tensorboard support

---------

+++ wisp/ops/raygen/raygen.py
@@ -15,8 +15,8 @@ from wisp.core import Rays
 # -- Supersample / Ray jitter --
 
 def generate_default_grid(width, height, device=None):
-    h_coords = torch.arange(height, device=device)
-    w_coords = torch.arange(width, device=device)
+    h_coords = torch.arange(height, device=device, dtype=torch.long)
+    w_coords = torch.arange(width, device=device, dtype=torch.long)
     return torch.meshgrid(h_coords, w_coords)  # return pixel_y, pixel_x
